project/mooc/requirements.txt

django==2.2.17
mysqlclient
Pillow
django-pure-pagination